// src/models/campaign.model.ts
export interface Campaign {
  id: string;
  name: string;
  why: string;
  goals: string;
  method: string;
  quote: string;
  missionquote: string;
  whatFor: string;
  how: string;
  contact: string;
  descriptionone: string;
  descriptiontwo: string;
  descriptionthree: string;
  descriptionfour: string;
  gallery: string[];
  subdomain: string;
}